+++ src/app/App.test.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
 import { render, unmountComponentAtNode } from "react-dom";
-import { Home } from "./../../containers";
+import App from "./App";
 
 it("renders without crashing", () => {
   const div = document.createElement("div");
-  render(<Home user="Default User" />, div);
+  render(<App />, div);
   unmountComponentAtNode(div);
 });
