+++ src/app/Routes.test.tsx
@@ -2,8 +2,8 @@ import { mount } from "enzyme";
 import * as React from "react";
 import { MemoryRouter } from "react-router-dom";
 
-import { About, Home } from "./../../containers";
-import Routes from "./../../Routes";
+import { About, Home } from "./components";
+import Routes from "./Routes";
 
 describe("Routes", () => {
   it("default path should redirect to Home component", () => {
