+++ src/app/components/About/index.ts
@@ -1,2 +1,1 @@
-export { default as Home } from "./Home";
 export { default as About } from "./About";
