+++ src/app/components/Home/Home.test.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
 import { render, unmountComponentAtNode } from "react-dom";
-import { About } from "./../../containers";
+import Home from "./Home";
 
 it("renders without crashing", () => {
   const div = document.createElement("div");
-  render(<About />, div);
+  render(<Home user="Default User" />, div);
   unmountComponentAtNode(div);
 });
