+++ src/app/components/index.ts
@@ -1,0 +1,2 @@
+export { About } from "./About";
+export { Home } from "./Home";
